+++ controlplane/src/core/repositories/analytics/SubgraphMetricsRepository.ts
@@ -97,7 +97,7 @@ export class SubgraphMetricsRepository {
       };
       return this.client.queryPromise<{ value: number | null }>(
         `
-        SELECT round(sum(total) / {multiplier:UInt32}, 4) AS value FROM (
+        SELECT round(sum(total) / {multiplier:Float64}, 4) AS value FROM (
         SELECT
           toDateTime({startDate:UInt32}) AS startDate,
           toDateTime({endDate:UInt32}) AS endDate,
@@ -131,7 +131,7 @@ export class SubgraphMetricsRepository {
       WITH
         toDateTime({startDate:UInt32}) AS startDate,
         toDateTime({endDate:UInt32}) AS endDate
-      SELECT hash, name, isPersisted, round(sum(total) / {multiplier:UInt32}, 4) AS value FROM (
+      SELECT hash, name, isPersisted, round(sum(total) / {multiplier:Float64}, 4) AS value FROM (
         SELECT
           Timestamp as timestamp,
           OperationHash as hash,
